notifications/pyproject.toml
[project]
name = "testing-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "niquests>=3.18.4",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.2.2",
]
